requirements.txt
PyPDF2>=3.0.0,<4.0.0
pandas>=1.5.0,<2.0.0
nltk>=3.8.1,<4.0.0
tqdm>=4.65.0,<5.0.0
python-dotenv>=1.0.0,<2.0.0
pyyaml>=6.0.0,<7.0.0
sentence-transformers>=2.2.2
transformers>=4.36.0
torch>=2.0.0
accelerate>=0.24.0
numpy>=1.24.0
faiss-cpu>=1.7.4
tiktoken>=0.5.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-chroma>=0.0.1
chromadb>=0.4.22
langchain-huggingface>=0.0.2 